--- source/Application/Transactions/AggregatedMeasureData/Notifications/NotifyWholesaleWhenAggregatedMeasureProcessIsInitialized.cs
@@ -1,39 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Application.Configuration.Commands;
-using Application.Transactions.AggregatedMeasureData.Commands;
-using Domain.Transactions.AggregatedMeasureData.Events;
-using MediatR;
-
-namespace Application.Transactions.AggregatedMeasureData.Notifications;
-
-public class NotifyWholesaleWhenAggregatedMeasureProcessIsInitialized : INotificationHandler<AggregatedMeasureProcessIsInitialized>
-{
-    private readonly ICommandScheduler _commandScheduler;
-
-    public NotifyWholesaleWhenAggregatedMeasureProcessIsInitialized(ICommandScheduler commandScheduler)
-    {
-        _commandScheduler = commandScheduler;
-    }
-
-    public async Task Handle(AggregatedMeasureProcessIsInitialized notification, CancellationToken cancellationToken)
-    {
-        if (notification == null) throw new ArgumentNullException(nameof(notification));
-        await _commandScheduler.EnqueueAsync(new NotifyWholesaleOfAggregatedMeasureDataRequest(notification.ProcessId.Id)).ConfigureAwait(false);
-    }
-}
