+++ source/Domain/Transactions/AggregatedMeasureData/Period.cs
@@ -17,20 +17,4 @@ using NodaTime;
 
 namespace Domain.Transactions.AggregatedMeasureData;
 
-public record Period(Instant Start, Instant End)
-{
-    public string StartToString()
-    {
-        return ParsePeriodDateFrom(Start);
-    }
-
-    public string EndToString()
-    {
-        return ParsePeriodDateFrom(End);
-    }
-
-    private static string ParsePeriodDateFrom(Instant instant)
-    {
-        return instant.ToString("yyyy-MM-ddTHH:mm'Z'", CultureInfo.InvariantCulture);
-    }
-}
+public record Period(Instant Start, Instant End);
