+++ source/Domain/Transactions/AggregatedMeasureData/ProcessEvents/AggregatedMeasureProcessWasAccepted.cs
@@ -14,7 +14,7 @@
 
 using Domain.SeedWork;
 
-namespace Domain.Transactions.AggregatedMeasureData.Events;
+namespace Domain.Transactions.AggregatedMeasureData.ProcessEvents;
 
 public class AggregatedMeasureProcessWasAccepted : DomainEvent
 {
