--- source/Infrastructure/Transactions/AggregatedMeasureData/Handlers/AcceptedAggregatedTimeSeriesFromWholesale.cs
@@ -1,52 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Application.Transactions.AggregatedMeasureData.Commands;
-using Domain.Transactions;
-using Domain.Transactions.AggregatedMeasureData;
-using Domain.Transactions.Exceptions;
-using Infrastructure.Configuration.Serialization;
-using MediatR;
-
-namespace Infrastructure.Transactions.AggregatedMeasureData.Handlers;
-
-public class AcceptedAggregatedTimeSeriesFromWholesale : IRequestHandler<AcceptedAggregatedTimeSeries, Unit>
-{
-    private readonly IAggregatedMeasureDataProcessRepository _aggregatedMeasureDataProcessRepository;
-    private readonly ISerializer _serializer;
-
-    public AcceptedAggregatedTimeSeriesFromWholesale(
-        IAggregatedMeasureDataProcessRepository aggregatedMeasureDataProcessRepository,
-        ISerializer serializer)
-    {
-        _aggregatedMeasureDataProcessRepository = aggregatedMeasureDataProcessRepository;
-        _serializer = serializer;
-    }
-
-    public async Task<Unit> Handle(AcceptedAggregatedTimeSeries request, CancellationToken cancellationToken)
-    {
-        ArgumentNullException.ThrowIfNull(request);
-
-        var process = await _aggregatedMeasureDataProcessRepository
-                          .GetByIdAsync(ProcessId.Create(request.ProcessId), cancellationToken).ConfigureAwait(false)
-                      ?? throw ProcessNotFoundException.ProcessForProcessIdNotFound(request.ProcessId);
-
-        process.WasAccepted(_serializer.Serialize(request.NotificationAggregatedTimeSerie));
-
-        return Unit.Value;
-    }
-}
