+++ source/Infrastructure/Transactions/AggregatedMeasureData/Notifications/Handlers/NotifyWholesaleWhenAggregatedMeasureProcessIsInitialized.cs
@@ -16,22 +16,22 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Application.Configuration.Commands;
-using Application.Transactions.AggregatedMeasureData.Commands;
-using Domain.Transactions.AggregatedMeasureData.Events;
+using Domain.Transactions.AggregatedMeasureData.ProcessEvents;
+using Infrastructure.Transactions.AggregatedMeasureData.Commands;
 using MediatR;
 
-namespace Application.Transactions.AggregatedMeasureData.Notifications;
+namespace Infrastructure.Transactions.AggregatedMeasureData.Notifications;
 
-public class SendAggregatedMeasureRequestToWholesaleWhenProcessIsSending : INotificationHandler<AggregatedMeasureProcessIsSending>
+public class NotifyWholesaleWhenAggregatedMeasureProcessIsInitialized : INotificationHandler<AggregatedMeasureProcessIsInitialized>
 {
     private readonly ICommandScheduler _commandScheduler;
 
-    public SendAggregatedMeasureRequestToWholesaleWhenProcessIsSending(ICommandScheduler commandScheduler)
+    public NotifyWholesaleWhenAggregatedMeasureProcessIsInitialized(ICommandScheduler commandScheduler)
     {
         _commandScheduler = commandScheduler;
     }
 
-    public async Task Handle(AggregatedMeasureProcessIsSending notification, CancellationToken cancellationToken)
+    public async Task Handle(AggregatedMeasureProcessIsInitialized notification, CancellationToken cancellationToken)
     {
         if (notification == null) throw new ArgumentNullException(nameof(notification));
         await _commandScheduler.EnqueueAsync(new SendAggregatedMeasureRequestToWholesale(notification.ProcessId.Id)).ConfigureAwait(false);
