+++ source/Infrastructure/Transactions/AggregatedMeasureData/Notifications/Handlers/WhenAnAcceptedAggregatedTimeSeriesRequestIsAvailable.cs
@@ -16,10 +16,10 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Application.Configuration.Commands;
-using Application.Transactions.AggregatedMeasureData.Commands;
+using Infrastructure.Transactions.AggregatedMeasureData.Commands;
 using MediatR;
 
-namespace Application.Transactions.AggregatedMeasureData.Notifications;
+namespace Infrastructure.Transactions.AggregatedMeasureData.Notifications.Handlers;
 
 public class WhenAnAcceptedAggregatedTimeSeriesRequestIsAvailable : INotificationHandler<AggregatedTimeSeriesRequestWasAccepted>
 {
